[core] Fix `test_object_spilling.py` on Windows (#53851)
I made the workload more stressful in
https://github.com/ray-project/ray/pull/53803 by fetching all of the
results concurrently. That seems to have caused Windows to time out:
https://buildkite.com/ray-project/postmerge/builds/10876#01977755-755b-485e-bd13-f0ea3e33cc36/158-818

I won't pretend to fully understand why, but reverting to the old
pattern in an attempt to fix it.

Also added an explicit wait for the dir to drain because there was an
error during cleanup caused by it.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>